Move login widgets to widgets folder
