+++ lib/widgets/login/options.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+
+import 'package:relabs_example/widgets/login/remember_checkbox.dart';
+
+class Options extends StatelessWidget {
+  const Options({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        const Row(
+          children: [
+            RememberCheckbox(),
+            SizedBox(width: 5),
+            Text(
+              'Remember-me',
+              style: TextStyle(
+                color: Colors.white,
+                fontWeight: FontWeight.w400,
+              ),
+            ),
+          ],
+        ),
+        Row(
+          children: [
+            TextButton(
+              onPressed: () { },
+              style: TextButton.styleFrom(
+                minimumSize: Size.zero,
+                padding: EdgeInsets.zero,
+              ),
+              child: const Text(
+                'Password Recovery',
+                style: TextStyle(
+                  color: Colors.white,
+                  fontWeight: FontWeight.w400,
+                ),
+              ),
+            ),
+          ],
+        )
+      ]
+    );
+  }
+}
